@@ -1,13 +1,12 @@
 package com.alexnassif.mobile.workoutbywill.Adapters
 
 import android.content.Context
-import android.support.v7.widget.RecyclerView
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
-import android.view.animation.AnimationUtils
 import android.widget.ImageView
 import android.widget.TextView
+import androidx.recyclerview.widget.RecyclerView
 import com.alexnassif.mobile.workoutbywill.Model.Exercise
 import com.alexnassif.mobile.workoutbywill.R
 import com.bumptech.glide.Glide
@@ -17,7 +16,7 @@ import com.bumptech.glide.Glide
  */
 class ExerciseAdapter(val context: Context, var exercises: List<Exercise>, val itemClick: (Exercise) -> Unit)
     : RecyclerView.Adapter<ExerciseAdapter.ExerciseHolder>() {
-    override fun onCreateViewHolder(parent: ViewGroup?, viewType: Int): ExerciseHolder {
+    override fun onCreateViewHolder(parent: ViewGroup, viewType: Int): ExerciseHolder {
         val view = LayoutInflater.from(context).inflate(R.layout.ex_type_cell_grid, parent, false)
         return ExerciseHolder(view, itemClick)
     }
@@ -35,9 +34,9 @@ class ExerciseAdapter(val context: Context, var exercises: List<Exercise>, val i
         return exercises.count()
     }
 
-    inner class ExerciseHolder(v: View?, val itemClick: (Exercise) -> Unit): RecyclerView.ViewHolder(v){
-        private val exerciseName = v?.findViewById<TextView>(R.id.exerciseTextView)
-        private val exerciseImage = v?.findViewById<ImageView>(R.id.exerciseImageView)
+    inner class ExerciseHolder(v: View, val itemClick: (Exercise) -> Unit): RecyclerView.ViewHolder(v){
+        private val exerciseName = v.findViewById<TextView>(R.id.exerciseTextView)
+        private val exerciseImage = v.findViewById<ImageView>(R.id.exerciseImageView)
 
         fun bindExercise(exercise: Exercise, context: Context){
             exerciseName?.text = exercise.name
