@@ -1,23 +1,19 @@
 package com.alexnassif.mobile.workoutbywill.Controller
 
 
-import android.arch.lifecycle.Observer
-import android.arch.lifecycle.ViewModelProviders
 import android.os.Bundle
-import android.support.v4.app.Fragment
-import android.support.v7.widget.GridLayoutManager
-import android.support.v7.widget.LinearLayoutManager
-import android.support.v7.widget.RecyclerView
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
+import androidx.fragment.app.Fragment
+import androidx.lifecycle.Observer
+import androidx.lifecycle.ViewModelProviders
+import androidx.recyclerview.widget.GridLayoutManager
 import com.alexnassif.mobile.workoutbywill.Adapters.WorkoutRecyclerAdapter
-import com.alexnassif.mobile.workoutbywill.Model.Workout
-
 import com.alexnassif.mobile.workoutbywill.R
-import com.alexnassif.mobile.workoutbywill.Services.DataService
+import com.alexnassif.mobile.workoutbywill.Utilities.InjectUtils
 import com.alexnassif.mobile.workoutbywill.ViewModel.WorkoutViewModel
-import com.google.firebase.auth.FirebaseAuth
+import com.alexnassif.mobile.workoutbywill.ViewModel.WorkoutViewModelFactory
 import kotlinx.android.synthetic.main.fragment_workout.*
 
 
@@ -34,7 +30,6 @@ class WorkoutFragment : Fragment() {
 
     override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
         super.onViewCreated(view, savedInstanceState)
-
 
         workoutRecyclerView.layoutManager = GridLayoutManager(context, 2)
 
@@ -55,8 +50,8 @@ class WorkoutFragment : Fragment() {
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
-
-        viewModel = ViewModelProviders.of(this).get(WorkoutViewModel::class.java)
+        val exerciseViewModelFactory: WorkoutViewModelFactory = InjectUtils.provideExerciseViewModelFactory(context!!.applicationContext)
+        viewModel = ViewModelProviders.of(this, exerciseViewModelFactory).get(WorkoutViewModel::class.java)
 
     }
 
