@@ -1,13 +1,12 @@
 package com.alexnassif.mobile.workoutbywill
 
 import android.content.Context
-import android.support.v7.widget.RecyclerView
 import android.util.AttributeSet
-import android.support.v7.widget.GridLayoutManager
-import android.util.Log
 import android.view.View
-import android.view.animation.GridLayoutAnimationController
 import android.view.ViewGroup
+import android.view.animation.GridLayoutAnimationController
+import androidx.recyclerview.widget.GridLayoutManager
+import androidx.recyclerview.widget.RecyclerView
 
 
 class GridRecyclerView : RecyclerView {
@@ -24,7 +23,7 @@ class GridRecyclerView : RecyclerView {
      */
     constructor(context: Context, attrs: AttributeSet, defStyle: Int) : super(context, attrs, defStyle) {}
 
-    override fun setLayoutManager(layout: RecyclerView.LayoutManager) {
+    override fun setLayoutManager(layout: RecyclerView.LayoutManager?) {
         if (layout is GridLayoutManager) {
             super.setLayoutManager(layout)
         } else {
