@@ -1,3 +1,3 @@
 package com.alexnassif.mobile.workoutbywill.Model
 
-class Program(val url: String, val name: String, val weeks: String)
+data class Program(val url: String, val name: String, val weeks: String)
