@@ -3,6 +3,6 @@ package com.alexnassif.mobile.workoutbywill.Model
 /**
  * Created by raymond on 1/27/18.
  */
-class Workout(val name: String) {
+data class Workout(val name: String) {
 
 }
