@@ -1,9 +1,8 @@
 package com.alexnassif.mobile.workoutbywill.ViewModel
 
-import android.arch.lifecycle.MutableLiveData
-import android.arch.lifecycle.ViewModel
+import androidx.lifecycle.MutableLiveData
+import androidx.lifecycle.ViewModel
 import com.alexnassif.mobile.workoutbywill.Model.Exercise
-import com.alexnassif.mobile.workoutbywill.Model.Workout
 import com.alexnassif.mobile.workoutbywill.Services.DataService
 
 class ExerciseViewModel: ViewModel() {
